requirements.txt
numpy
tensorflow==0.12
tqdm
sklearn
scipy
riseml
colour
pillow

